--- auth/src/main/java/com/android/swingmusic/auth/presentation/navigation/AuthNavigator.kt
@@ -1,13 +1,0 @@
-package com.android.swingmusic.auth.presentation.navigation
-
-interface AuthNavigator {
-
-    fun gotoLoginWithUsername()
-
-    fun gotoLoginWithQrCode()
-
-    fun gotoHomeNavGraph()
-
-    // Todo: Remove this after adding home content
-    fun gotoFolderNavGraph()
-}
