--- feature/album/src/main/java/com/android/swingmusic/album/presentation/navigator/AlbumNavigator.kt
@@ -1,8 +1,0 @@
-package com.android.swingmusic.album.presentation.navigator
-
-interface AlbumNavigator {
-
-    fun gotoAlbumWithInfo(albumHash: String)
-
-    fun navigateBack()
-}
