--- feature/artist/src/main/java/com/android/swingmusic/artist/presentation/navigator/ArtistNavigator.kt
@@ -1,5 +1,0 @@
-package com.android.swingmusic.artist.presentation.navigator
-
-interface ArtistNavigator {
-    fun gotoArtistInfo(artistHash: String)
-}
