--- feature/player/src/main/java/com/android/swingmusic/player/presentation/navigator/PlayerNavigator.kt
@@ -1,7 +1,0 @@
-package com.android.swingmusic.player.presentation.navigator
-
-interface PlayerNavigator {
-    fun navigateBack()
-
-    fun gotoQueueScreen()
-}
