+++ feature/player/src/main/java/com/android/swingmusic/player/presentation/screen/NowPlaying.kt
@@ -54,13 +54,13 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
+import com.android.swingmusic.common.presentation.navigator.CommonNavigator
 import com.android.swingmusic.core.domain.model.Track
 import com.android.swingmusic.core.domain.model.TrackArtist
 import com.android.swingmusic.core.domain.util.PlaybackState
 import com.android.swingmusic.core.domain.util.RepeatMode
 import com.android.swingmusic.core.domain.util.ShuffleMode
 import com.android.swingmusic.player.presentation.event.PlayerUiEvent
-import com.android.swingmusic.player.presentation.navigator.PlayerNavigator
 import com.android.swingmusic.player.presentation.viewmodel.MediaControllerViewModel
 import com.android.swingmusic.uicomponent.R
 import com.android.swingmusic.uicomponent.presentation.component.slider.WaveAnimationSpecs
@@ -542,7 +542,7 @@ private fun NowPlaying(
 @Composable
 fun NowPlayingScreen(
     mediaControllerViewModel: MediaControllerViewModel,
-    navigator: PlayerNavigator
+    navigator: CommonNavigator
 ) {
     val playerUiState by mediaControllerViewModel.playerUiState.collectAsState()
     val baseUrl by mediaControllerViewModel.baseUrl.collectAsState()
